chore: 🤖 Updated the test docker scripts
