+++ scripts/docker-compose.yml
@@ -4,13 +4,12 @@ services:
     image: 'gitlab/gitlab-ce'
     container_name: 'gitlab'
     environment:
-      GITLAB_URL: 'http://localhost:8080'
-      PERSONAL_ACCESS_TOKEN: ""
+      PERSONAL_ACCESS_TOKEN: ''
       GITLAB_OMNIBUS_CONFIG: |
-          gitlab_rails['monitoring_whitelist'] = ['0.0.0.0/0', '172.17.0.1'];
+        gitlab_rails['monitoring_whitelist'] = ['0.0.0.0/0', '172.17.0.1'];
       GITLAB_ROOT_PASSWORD: 'password'
     ports:
       - '8080:80'
       - '8443:443'
     volumes:
-      - './init_data.rb:/tmp/init_data.rb'
+      - './init.rb:/mnt/init.rb'
