+++ scripts/init.rb
@@ -1,0 +1,9 @@
+#!/usr/bin/env ruby
+u = User.first
+u.admin = true
+u.save!
+
+t = PersonalAccessToken.new({ user: u, name: 'gitbeaker', scopes: ['api', 'read_user']}) 
+t.save! 
+ 
+puts t.token
