--- test/docker/init_data.rb
@@ -1,7 +1,0 @@
-#!/usr/bin/env ruby
-u = User.first
-
-t = PersonalAccessToken.new({ user: u, name: 'node-gitlab', scopes: ['api']}) 
-t.save! 
- 
-puts t.token
